requirements.txt
Flask==3.0.0
requests==2.31.0
python-dotenv==1.0.0
flask-restful==0.3.10
flask-cors==4.0.0
pymongo==4.6.1
apscheduler==3.10.4
